Some updates

+++ backup.py
@@ -5,26 +5,35 @@ import MySQLdb
 import hashlib
 import datetime
 import ConfigParser
+import binascii
+import socket
 from time import sleep
 from subprocess import Popen, PIPE
 
 
+def ip_to_hex(ip):
+    return binascii.hexlify(socket.inet_aton(ip)).upper()
+
+
 def write_in_log(wil_message, wil_ip=''):
 
-    date_time = str(datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%d"))
-
     try:
-        wil_log_file = open(path_to_log_file, 'a')
+        wil_log_file = open(path_to_log_file, 'a', buffering=-1)
     except IOError as e:
         print "Can not open log file. I/O Error({0}): {1}".format(e.errno, e.strerror)
         exit(1)
     else:
+        date_time = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
         if wil_ip != '':
-            wil_result_message = date_time + " " + wil_ip + " " + wil_message + '\n'
+            wil_result_message = date_time + " " + wil_ip + " " + wil_message
         else:
-            wil_result_message = date_time + " " + wil_message + '\n'
+            wil_result_message = date_time + " " + wil_message
 
-        wil_log_file.write(wil_result_message)
+        print wil_result_message
+
+        del date_time
+
+        wil_log_file.write(wil_result_message + '\n')
         wil_log_file.close()
 
 
@@ -39,12 +48,12 @@ def do_backup_config(dbc_ip, dbc_community, dbc_type, dbc_tftp, dbc_file_name):
         # 25 - D-link DES-3200-18 hw A1/B1
         # 41 - D-link DES-3200-26 hw A1/B1
 
-        Popen("snmpset -v2c -c " + dbc_community + " " + dbc_ip + " 1.3.6.1.4.1.171.12.1.2.1.1.3.3 a " + dbc_tftp + " \
-            1.3.6.1.4.1.171.12.1.2.1.1.4.3 i 2 \
-            1.3.6.1.4.1.171.12.1.2.1.1.5.3 s dlink/" + dbc_file_name + " \
-            1.3.6.1.4.1.171.12.1.2.1.1.7.3 i 2 \
-            1.3.6.1.4.1.171.12.1.2.1.1.8.3 i 3",
-              shell=True, stdin=PIPE, stdout=PIPE).stdout.read().split()
+        snmp_command = "snmpset -v2c -c " + dbc_community + " " + dbc_ip + " \
+                    1.3.6.1.4.1.171.12.1.2.1.1.3.3 a " + dbc_tftp + " \
+                    1.3.6.1.4.1.171.12.1.2.1.1.4.3 i 2 \
+                    1.3.6.1.4.1.171.12.1.2.1.1.5.3 s dlink/" + dbc_file_name + " \
+                    1.3.6.1.4.1.171.12.1.2.1.1.7.3 i 2 \
+                    1.3.6.1.4.1.171.12.1.2.1.1.8.3 i 3"
 
     elif dbc_type in [27, 28, 29, 30, 32, 33, 39, 40]:
 
@@ -58,38 +67,47 @@ def do_backup_config(dbc_ip, dbc_community, dbc_type, dbc_tftp, dbc_file_name):
         # 39 - D-link DES-3200-28 hw C1
         # 40 - D-link DES-3200-18 hw C1
 
-        Popen("snmpset -v2c -c " + dbc_community + " " + dbc_ip + " 1.3.6.1.4.1.171.12.1.2.18.1.1.3.3 a " + dbc_tftp + " \
-            1.3.6.1.4.1.171.12.1.2.18.1.1.5.3 s dlink/" + dbc_file_name + " \
-            1.3.6.1.4.1.171.12.1.2.18.1.1.8.3 i 2 \
-            1.3.6.1.4.1.171.12.1.2.18.1.1.12.3 i 3",
-              shell=True, stdin=PIPE, stdout=PIPE).stdout.read().split()
+        snmp_command = "snmpset -v2c -c " + dbc_community + " " + dbc_ip + " \
+                    1.3.6.1.4.1.171.12.1.2.18.1.1.3.3 a " + dbc_tftp + " \
+                    1.3.6.1.4.1.171.12.1.2.18.1.1.5.3 s dlink/" + dbc_file_name + " \
+                    1.3.6.1.4.1.171.12.1.2.18.1.1.8.3 i 2 \
+                    1.3.6.1.4.1.171.12.1.2.18.1.1.12.3 i 3"
 
     elif dbc_type in [35]:
 
         # Type of support switches:
         # 35 - D-link DGS-1100-10/ME
 
-        Popen("snmpset -v2c -c " + dbc_community + " " + dbc_ip + " 1.3.6.1.4.1.171.10.134.2.1.3.2.1.0 x C0A810B4 \
+        snmp_command = "snmpset -v2c -c " + dbc_community + " " + dbc_ip + " \
+                    1.3.6.1.4.1.171.10.134.2.1.3.2.1.0 x " + ip_to_hex(dbc_tftp) + " \
                     1.3.6.1.4.1.171.10.134.2.1.3.2.2.0 i 1 \
                     1.3.6.1.4.1.171.10.134.2.1.3.2.4.0 s dlink/" + dbc_file_name + " \
-                    1.3.6.1.4.1.171.10.134.2.1.3.2.5.0 i 2",
-              shell=True, stdin=PIPE, stdout=PIPE).stdout.read().split()
+                    1.3.6.1.4.1.171.10.134.2.1.3.2.5.0 i 2"
 
 #    elif dbc_type in [37]:
 
         # Type of support switches:
         # 35 - D-link DGS-1100-10/ME
 
-#        Popen("snmpset -v2c -c " + dbc_community + " " + dbc_ip + " 1.3.6.1.4.1.171.10.139.3.1.1.1.2.4.2.0 x C0A810B4 \
-#                    1.3.6.1.4.1.171.10.139.3.1.1.1.2.4.3.0 s dlink/" + dbc_file_name + " \
-#                    1.3.6.1.4.1.171.10.139.3.1.1.1.2.4.4.0 i 1 \
-#                    1.3.6.1.4.1.171.10.139.3.1.1.1.2.4.6.0 i 2",
+#        Popen("snmpset -v2c -c " + dbc_community + " " + dbc_ip + " \
+#                   1.3.6.1.4.1.171.10.139.3.1.1.1.2.4.2.0 x C0A810B4 \
+#                   1.3.6.1.4.1.171.10.139.3.1.1.1.2.4.3.0 s dlink/" + dbc_file_name + " \
+#                   1.3.6.1.4.1.171.10.139.3.1.1.1.2.4.4.0 i 1 \
+#                   1.3.6.1.4.1.171.10.139.3.1.1.1.2.4.6.0 i 2",
 #              shell=True, stdin=PIPE, stdout=PIPE).stdout.read().split()
 
     else:
 
+        snmp_command = ''
+
         write_in_log("I do not know how to do backup from this switch", dbc_ip)
         print "I do not know how to do backup from this switch. " + dbc_ip
+
+    if snmp_command != '':
+        Popen(snmp_command, shell=True, stdin=PIPE, stdout=PIPE).stdout.read().split()
+        return True
+    else:
+        return False
 
 
 def get_md5_sum(gms_file_name):
@@ -103,7 +121,8 @@ def get_md5_sum(gms_file_name):
 
 def check_config(cc_device_id, cc_hash):
     cursor3 = db.cursor()
-    cursor3.execute("SELECT hash FROM backup WHERE `devid` = '" + cc_device_id + "' ORDER BY date DESC LIMIT 1")
+    cursor3.execute("SELECT hash FROM `" + table_name + "` WHERE `devid` = '" + cc_device_id + "'\
+                    ORDER BY date DESC LIMIT 1")
     for cc_row in cursor3.fetchall():
         if cc_row[0] == cc_hash:
             return True
@@ -139,6 +158,7 @@ host_name = config.get('database', 'host')
 user_name = config.get('database', 'user')
 password = config.get('database', 'passwd')
 db_name = config.get('database', 'db')
+table_name = config.get('database', 'table')
 
 db = MySQLdb.connect(host=host_name,
                      user=user_name,
@@ -153,7 +173,7 @@ cursor.execute("SELECT `ip`, `access_snmp_write`, `devices`.`type`, `id` \
                     ON `devices`.`type` = `devices_config`.`type` \
                 WHERE `ping` = '1' \
                     AND `devices_config`.`do_backup` = '1' \
-                LIMIT 100")
+                LIMIT 15")
 
 for row in cursor.fetchall():
 
@@ -163,27 +183,22 @@ for row in cursor.fetchall():
     device_id = str(row[3])
 
     file_name = ip_address + "_" + str(datetime.date.today()) + "_" + get_random_word() + ".cfg"
-    do_backup_config(ip_address, snmp_community, device_type, ip_tftp_server, file_name)
+    if do_backup_config(ip_address, snmp_community, device_type, ip_tftp_server, file_name):
+        sleep(5)
+        if os.path.isfile(path_to_tftp_folder + file_name):
+            md5_hash = str(get_md5_sum(path_to_tftp_folder + file_name))
 
-    sleep(5)
-
-    if os.path.isfile(path_to_tftp_folder + file_name):
-        md5_hash = str(get_md5_sum(path_to_tftp_folder + file_name))
-
-        if check_config(device_id, md5_hash):
-            print device_id + " Config file was not updated in last time"
-            write_in_log("Config file was not updated in last time", ip_address)
-            remove_file(file_name)
+            if check_config(device_id, md5_hash):
+                write_in_log("Config file was not updated in last time", ip_address)
+                remove_file(file_name)
+            else:
+                move_file_to_archive(file_name)
+                cursor.execute("insert into `" + table_name + "` (devid, fname, hash)\
+                                VALUES(" + device_id + ", '" + file_name + "', '" + md5_hash + "')")
+                db.commit()
+                write_in_log("Config was updated...", ip_address)
         else:
-            move_file_to_archive(file_name)
-            cursor.execute("insert into backup (devid, fname, hash)\
-                            VALUES(" + device_id + ", '" + file_name + "', '" + md5_hash + "')")
-            db.commit()
-            print "Config was updated..."
-            write_in_log("Config was updated...", ip_address)
-    else:
-        print "File not found"
-        write_in_log("File not found", device_id)
+            write_in_log("File not found", device_id)
 
 db.commit()
 db.close()
